gjorde ferdig cost func, og prøvde å gjøre litt endringer for å unngå sirkulær import
