--- datatypes/hardware_types.go
@@ -1,1 +1,0 @@
-package datatypes
