--- requests/request_assigner.go
@@ -1,3 +1,0 @@
-package requests
-
-// skal basert på en global hall calls liste (?) assigne hall calls til best egnede heis, vha en cost function
