+++ requests/request_utilities.go
@@ -1,11 +1,11 @@
 package requests
 
 import (
-	"project/elevio"
 	"project/elevator_control"
+	"project/elevio"
 )
 
-func RequestsAbove(elevator *Elevator) bool { // skal returnere true/false om det er noen aktive orders i etasjer over
+func RequestsAbove(elevator *elevator_control.Elevator) bool { // skal returnere true/false om det er noen aktive orders i etasjer over
 	for f := elevator.CurrentFloor + 1; f < len(elevator.Orders); f++ {
 		for _, order := range elevator.Orders[f] {
 			if order {
@@ -16,7 +16,7 @@ func RequestsAbove(elevator *Elevator) bool { // skal returnere true/false om de
 	return false
 }
 
-func RequestsBelow(elevator *Elevator) bool { // skal returnere true/false om det er noen aktive orders i etasjer under
+func RequestsBelow(elevator *elevator_control.Elevator) bool { // skal returnere true/false om det er noen aktive orders i etasjer under
 	for f := elevator.CurrentFloor - 1; f >= 0; f-- {
 		for _, order := range elevator.Orders[f] {
 			if order {
@@ -27,7 +27,7 @@ func RequestsBelow(elevator *Elevator) bool { // skal returnere true/false om de
 	return false
 }
 
-func RequestsHere(elevator *Elevator) bool {
+func RequestsHere(elevator *elevator_control.Elevator) bool {
 	for _, order := range elevator.Orders[elevator.CurrentFloor] {
 		if order {
 			return true
@@ -36,7 +36,7 @@ func RequestsHere(elevator *Elevator) bool {
 	return false
 }
 
-func RequestsHereMatchingDir(elevator *Elevator, dir elevio.MotorDirection) bool {
+func RequestsHereMatchingDir(elevator *elevator_control.Elevator, dir elevio.MotorDirection) bool {
 	for b, active := range elevator.Orders[elevator.CurrentFloor] {
 		if !active {
 			continue
@@ -56,7 +56,7 @@ func RequestsHereMatchingDir(elevator *Elevator, dir elevio.MotorDirection) bool
 	return false
 }
 
-func AddOrder(elevator *Elevator, floor int, button elevio.ButtonType) {
+func AddOrder(elevator *elevator_control.Elevator, floor int, button elevio.ButtonType) {
 	elevator.Orders[floor][button] = true
 	elevio.SetButtonLamp(button, floor, true)
 
@@ -72,7 +72,7 @@ func AddOrder(elevator *Elevator, floor int, button elevio.ButtonType) {
 //}
 
 // Ny ClearRequestsAtFloor som kun  sletter om retningen om den stemmer med kjøreretning, og cab calls
-func ClearRequestsAtFloor(elevator *Elevator) {
+func ClearRequestsAtFloor(elevator *elevator_control.Elevator) {
 	floor := elevator.CurrentFloor
 	for b := 0; b < 3; b++ {
 		if !elevator.Orders[floor][b] {
@@ -98,7 +98,7 @@ func ClearRequestsAtFloor(elevator *Elevator) {
 	}
 }
 
-func ClearAllRequests(elevator *Elevator) {
+func ClearAllRequests(elevator *elevator_control.Elevator) {
 	for f := 0; f < len(elevator.Orders); f++ {
 		for b := 0; b < 3; b++ {
 			elevator.Orders[f][b] = false
@@ -107,7 +107,7 @@ func ClearAllRequests(elevator *Elevator) {
 	}
 }
 
-func ChooseDirection(elevator *Elevator) elevio.MotorDirection { //velger retning basert på nåværende retning og bestillinger
+func ChooseDirection(elevator *elevator_control.Elevator) elevio.MotorDirection { //velger retning basert på nåværende retning og bestillinger
 	switch elevator.Direction {
 	case elevio.MD_Up:
 		if RequestsAbove(elevator) {
@@ -135,7 +135,7 @@ func ChooseDirection(elevator *Elevator) elevio.MotorDirection { //velger retnin
 
 }
 
-func ShouldStop(elevator *Elevator) bool {
+func ShouldStop(elevator *elevator_control.Elevator) bool {
 	switch elevator.Direction {
 	case elevio.MD_Up:
 		return RequestsHereMatchingDir(elevator, elevio.MD_Up) || !RequestsAbove(elevator)
